+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Get the scheduled times by the CRON expression
 * `logError` now also accepts a throwable for the server log
 * Reworked system job `JobExecutionCleanupWorker` to clear old execution for all jobs in one execution
+* Dubious way to manually trigger `JobWorkerWith.scheduledJobExecutionCreation`: `JobEngineService.triggerScheduledJobExecutionCreation()`
 
 ### BREAKING CHANGES